+++ packages/vlossom/src/components/index.ts
@@ -18,6 +18,9 @@ export { default as VsCheckboxSet } from './vs-checkbox/VsCheckboxSet.vue';
 export type { VsChipStyleSet } from './vs-chip/types';
 export { default as VsChip } from './vs-chip/VsChip.vue';
 
+export type { ConfirmOptions, PromptOptions } from './vs-confirmation/types';
+export { default as VsConfirm } from './vs-confirmation/VsConfirm.vue';
+
 export { default as VsContainer } from './vs-container/VsContainer.vue';
 
 export type { VsDividerStyleSet } from './vs-divider/types';
