+++ packages/vlossom/src/components/vs-confirmation/types.ts
