+++ packages/vlossom/src/nodes/index.ts
@@ -1,6 +1,3 @@
-export type { ConfirmOptions, PromptOptions } from './vs-confirm/types';
-export { default as VsConfirmation } from './vs-confirm/VsConfirmation.vue';
-
 export { default as VsContentRenderer } from './vs-content-renderer/VsContentRenderer.vue';
 
 export type { VsModalNodeStyleSet, VsModalOptions } from './vs-modal-node/types';
